0-print_list.c

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -7,6 +7,7 @@
  *
  * Return: the number of nodes printed
  */
+
 size_t print_list(const list_t *h)
 {
 	size_t count = 0;
@@ -16,7 +17,8 @@ size_t print_list(const list_t *h)
 		if (h->str == NULL)
 			printf("[0] (nil)\n");
 		else
-			printf("[%d] %s\n", h->len, h->str);
+			printf("[%ld] %s\n", (long)h->len, h->str);
+
 		count++;
 		h = h->next;
 	}
